Python/v2-p/requirements.txt
fastapi  
httpx  
langchain  
langchain-google-genai  
langchain-groq  
langsmith  
llama-index  
llama-index-embeddings-huggingface  
llama-index-llms-groq
logging
mysql-connector-python  
oci  
pydantic-settings  
PyMuPDF  
pypdf2  
python-multipart  
uvicorn
sendgrid
langgraph

# search engine
ddgs

# Testes
pytest
pytest-asyncio

#PDF
pytesseract
pypdfium2[jpeg-turbo]
Pillow